Move map utils to core/
